requirements.txt
haystack-ai
fastapi
uvicorn[standard]
streamlit
pypdf
sentence-transformers
llama-cpp-python
PyMuPDF
python-dotenv
requests
haystack-chroma
